requirements.txt
pytest==5.2.2
pyspark==2.4.4
mock==3.0.5
numpy==1.17.3
pandas==0.25.2
boto3==1.10.13
s3fs==0.3.5
py4j==0.10.7
pyarrow==0.15.1
awsglue-local
awswrangler==1.0.4
#git+https://github.com/awslabs/aws-glue-libs.git#subdirectory=awsglue